pyproject.toml
[tool.poetry]
name = 'gopher_bot'
version = '0.1.0'
description = 'Simple server stats tracking bot with support for wh00t server chat'
authors = ['roboto84 <7658704+roboto84@users.noreply.github.com>']

[tool.poetry.dependencies]
python = '^3.8'
python-dotenv = '^0.15.0'
psutil = '^5.8.0'
wh00t-core = {git = 'https://github.com/roboto84/wh00t_core.git', tag = 'v0.4.1'}

[tool.poetry.dev-dependencies]
pytest = '^5.2'

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
